Add standard test suite infrastructure
Import all the standard module tests from Term::ANSIColor and other
places, and then update them for a Module::Build distribution.  Change
the copyright lines to match the rest of the package and refer to
LICENSE for licensing information.  Tweak t/strict.t to ignore the
generated Build script, which doesn't use strict.