Change the stream implementation to be lazy
Make use of a helper method in Log::Stream to process more data
only when we have to, rather than proactively.  This means that
both head() and get() will only retrieve data when they need to,
and processing of the first line is deferred to the first head()
or get() call.