Change the stream and dump output of webkdc-user-agent
Modify the merged stream to only return the minimum of data that
we care about: the timestamp, the user, and the user agent string.
Remove the -d flag and add a -c flag, for CSV, that outputs the
result stream as CSV.  Eventually we'll support loading data from
that output and using it to generate new reports.