Further optimizations and cleanups of Apache log parsing
Turn the timestamp parsing functions into static helper functions
and use a lexical variable for the cache instead of the object.
Catch errors when parsing the timestamp and interpret that as an
unparsable line.  Further optimize and clean up how the hash is
built in Log::Stream::Parse::Apache::Combined.